Rename SurveyDataChart to SurveyAnswerChart
