--- src/components/container/SurveyDataChart/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./SurveyDataChart";
