+++ src/components/container/index.ts
@@ -32,7 +32,7 @@ export { default as RelativeActivityDayNavigator } from "./RelativeActivityDayNa
 export { default as RelativeActivity } from "./RelativeActivity"
 export { default as RelativeActivityToday } from "./RelativeActivityToday"
 export { default as RestingHeartRateCalendar } from "./RestingHeartRateCalendar"
-export { default as SurveyDataChart } from "./SurveyDataChart"
+export { default as SurveyAnswerChart } from "./SurveyAnswerChart"
 export { default as SurveyTaskList } from "./SurveyTaskList"
 export { default as SeverityCalendar } from "./SeverityCalendar"
 export { default as ViewEhr } from "./ViewEhr"
